backend: allow negative timestamps in commits and operations
I was reading a draft of "Git Rev News: Edition 91" [1] where Peff
mentions some unfinished patches to allow negative timestamps in
Git. So I figured I should add support for that before I forget. I
haven't checked if libgit2 supports it, so it might be that our Git
backend still doesn't support it after this patch.

 [1] https://github.com/git/git.github.io/blob/master/rev_news/drafts/edition-91.md

+++ lib/src/backend.rs
@@ -226,7 +226,7 @@ pub enum Phase {
 }
 
 #[derive(Debug, PartialEq, Eq, Clone, PartialOrd, Ord)]
-pub struct MillisSinceEpoch(pub u64);
+pub struct MillisSinceEpoch(pub i64);
 
 #[derive(Debug, PartialEq, Eq, Clone, PartialOrd, Ord)]
 pub struct Timestamp {
@@ -244,7 +244,7 @@ impl Timestamp {
         datetime: chrono::DateTime<Tz>,
     ) -> Self {
         Self {
-            timestamp: MillisSinceEpoch(datetime.timestamp_millis() as u64),
+            timestamp: MillisSinceEpoch(datetime.timestamp_millis()),
             tz_offset: datetime.offset().local_minus_utc() / 60,
         }
     }
